src/third_party/blink/web_tests/serial/serialPort_readable_open.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/unguessable_token.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/serial/serial.mojom.js"></script>
<script src="resources/serial-test-utils.js"></script>
<script>

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);

  assert_equals(port.readable, null);

  await port.open({ baudrate: 9600 });
  const readable = port.readable;
  assert_true(readable instanceof ReadableStream);

  await port.close();
  assert_equals(port.readable, null);

  const reader = readable.getReader();
  const { value, done } = await reader.read();
  assert_true(done);
  assert_equals(value, undefined);
}, 'SerialPort.readable is set by open() and closes on port close');

</script>
